+++ example-apps/test/dashboard/server.js
@@ -1,7 +1,0 @@
-module.exports = {
-
-  "users.listed": () => {},
-
-  "users.created": () => {}
-
-}