+++ src/cmd/shared.go
@@ -5,12 +5,10 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"path"
 
 	"github.com/Originate/exosphere/src/aws"
 	"github.com/Originate/exosphere/src/docker/composebuilder"
 	"github.com/Originate/exosphere/src/types"
-	"github.com/Originate/exosphere/src/util"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
@@ -44,18 +42,13 @@ func getAwsConfig(profile string) (types.AwsConfig, error) {
 	if err != nil {
 		return types.AwsConfig{}, err
 	}
-	homeDir, err := util.GetHomeDirectory()
-	if err != nil {
-		return types.AwsConfig{}, fmt.Errorf("Cannot get home directory: %s", err)
-	}
 	return types.AwsConfig{
 		Region:               appConfig.Production.Region,
 		AccountID:            appConfig.Production.AccountID,
 		SslCertificateArn:    appConfig.Production.SslCertificateArn,
 		Profile:              profile,
-		CredentialsFile:      path.Join(homeDir, ".aws", "credentials"),
-		SecretsBucket:        fmt.Sprintf("%s-terraform-secrets", appConfig.Name),
-		TerraformStateBucket: fmt.Sprintf("%s-terraform", appConfig.Name),
+		SecretsBucket:        fmt.Sprintf("%s-%s-terraform-secrets", appConfig.Production.AccountID, appConfig.Name),
+		TerraformStateBucket: fmt.Sprintf("%s-%s-terraform", appConfig.Production.AccountID, appConfig.Name),
 		TerraformLockTable:   "TerraformLocks",
 	}, nil
 }