+++ src/terraform/template_builder.go
@@ -8,6 +8,7 @@ import (
 	"path/filepath"
 
 	"github.com/Originate/exosphere/src/types"
+	"github.com/Originate/exosphere/src/util"
 	"github.com/hoisie/mustache"
 	"github.com/pkg/errors"
 )
@@ -28,14 +29,31 @@ func WriteTerraformFile(data string, terraformDir string) error {
 	var filePerm os.FileMode = 0744 //standard Unix file permission: rwxrw-rw-
 	err := os.MkdirAll(terraformDir, filePerm)
 	if err != nil {
-		return errors.Wrap(err, "Failed to get create directory")
+		return errors.Wrap(err, "Failed to get create 'terraform' directory")
+	}
+
+	err = writeGitIgnore(terraformDir)
+	if err != nil {
+		return errors.Wrap(err, "Failed to write 'terraform/.gitignore' file")
 	}
 
 	err = ioutil.WriteFile(filepath.Join(terraformDir, terraformFile), []byte(data), filePerm)
 	if err != nil {
-		return errors.Wrap(err, "Failed writing Terraform files")
+		return errors.Wrap(err, "Failed to write 'terraform/main.tf' file")
 	}
 	return nil
+}
+
+func writeGitIgnore(terraformDir string) error {
+	gitIgnore := `.terraform/
+terraform.tfstate
+terraform.tfstate.backup`
+	gitIgnorePath := filepath.Join(terraformDir, ".gitignore")
+	fileExists, err := util.DoesFileExist(gitIgnorePath)
+	if !fileExists {
+		return ioutil.WriteFile(gitIgnorePath, []byte(gitIgnore), 0744)
+	}
+	return err
 }
 
 func getTemplate(template string) (string, error) {
@@ -49,8 +67,8 @@ func getTemplate(template string) (string, error) {
 // ReadTerraformFile reads the contents of the main terraform file
 func ReadTerraformFile(deployConfig types.DeployConfig) ([]byte, error) {
 	terraformFilePath := filepath.Join(deployConfig.TerraformDir, terraformFile)
-	var err error
-	if _, err = os.Stat(terraformFilePath); !os.IsNotExist(err) {
+	fileExists, err := util.DoesFileExist(terraformFilePath)
+	if fileExists {
 		return ioutil.ReadFile(terraformFilePath)
 	}
 	return []byte{}, err