+++ src/types/app_config.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"path"
+	"regexp"
 	"sort"
 
 	"github.com/Originate/exosphere/src/util"
@@ -32,7 +33,7 @@ func NewAppConfig(appDir string) (result AppConfig, err error) {
 	if err != nil {
 		return result, errors.Wrap(err, "Failed to unmarshal application.yml")
 	}
-	return result, nil
+	return result, result.validateAppConfig()
 }
 
 // GetDevelopmentDependencyNames returns the names of all dev dependencies listed in appConfig
@@ -53,30 +54,30 @@ func (a AppConfig) GetProductionDependencyNames() []string {
 	return result
 }
 
-// GetServiceData returns the configurations data for the given services
+// GetServiceData returns the configuration data listed under a service role in application.yml
 func (a AppConfig) GetServiceData() map[string]ServiceData {
 	result := make(map[string]ServiceData)
-	a.forEachService(func(serviceType, serviceName string, data ServiceData) {
-		result[serviceName] = data
+	a.forEachService(func(serviceType, serviceRole string, data ServiceData) {
+		result[serviceRole] = data
 	})
 	return result
 }
 
-// GetSortedServiceNames returns the service names for the given services sorted alphabetically
-func (a AppConfig) GetSortedServiceNames() []string {
+// GetSortedServiceRoles returns the service roles listed in application.yml sorted alphabetically
+func (a AppConfig) GetSortedServiceRoles() []string {
 	result := []string{}
-	a.forEachService(func(serviceType, serviceName string, data ServiceData) {
-		result = append(result, serviceName)
+	a.forEachService(func(serviceType, serviceRole string, data ServiceData) {
+		result = append(result, serviceRole)
 	})
 	sort.Strings(result)
 	return result
 }
 
-// GetServiceProtectionLevels returns a map containing service names to their protection level
+// GetServiceProtectionLevels returns a map containing service role to protection level
 func (a AppConfig) GetServiceProtectionLevels() map[string]string {
 	result := make(map[string]string)
-	a.forEachService(func(serviceType, serviceName string, data ServiceData) {
-		result[serviceName] = serviceType
+	a.forEachService(func(serviceType, serviceRole string, data ServiceData) {
+		result[serviceRole] = serviceType
 	})
 	return result
 }
@@ -93,35 +94,50 @@ func (a AppConfig) GetSilencedDevelopmentDependencyNames() []string {
 	return result
 }
 
-// GetSilencedServiceNames returns the names of services that are configured
+// GetSilencedServiceRoles returns the names of services that are configured
 // as silent
-func (a AppConfig) GetSilencedServiceNames() []string {
+func (a AppConfig) GetSilencedServiceRoles() []string {
 	result := []string{}
-	a.forEachService(func(serviceType, serviceName string, data ServiceData) {
+	a.forEachService(func(serviceType, serviceRole string, data ServiceData) {
 		if data.Silent {
-			result = append(result, serviceName)
+			result = append(result, serviceRole)
 		}
 	})
 	return result
 }
 
-// VerifyServiceDoesNotExist returns an error if the service serviceRole already
+// VerifyServiceRoleDoesNotExist returns an error if the serviceRole already
 // exists in existingServices, and return nil otherwise.
-func (a AppConfig) VerifyServiceDoesNotExist(serviceRole string) error {
-	if util.DoesStringArrayContain(a.GetSortedServiceNames(), serviceRole) {
-		return fmt.Errorf(`Service %v already exists in this application`, serviceRole)
+func (a AppConfig) VerifyServiceRoleDoesNotExist(serviceRole string) error {
+	if util.DoesStringArrayContain(a.GetSortedServiceRoles(), serviceRole) {
+		return fmt.Errorf(`Service role '%v' already exists in this application`, serviceRole)
 	}
 	return nil
 }
 
 func (a AppConfig) forEachService(fn func(string, string, ServiceData)) {
-	for serviceName, data := range a.Services.Worker {
-		fn("worker", serviceName, data)
+	for serviceRole, data := range a.Services.Worker {
+		fn("worker", serviceRole, data)
 	}
-	for serviceName, data := range a.Services.Private {
-		fn("private", serviceName, data)
+	for serviceRole, data := range a.Services.Private {
+		fn("private", serviceRole, data)
 	}
-	for serviceName, data := range a.Services.Public {
-		fn("public", serviceName, data)
+	for serviceRole, data := range a.Services.Public {
+		fn("public", serviceRole, data)
 	}
+}
+
+func (a AppConfig) validateAppConfig() error {
+	appNameRegex := regexp.MustCompile("^[a-z0-9]+(-[a-z0-9]+)*$")
+	if !appNameRegex.MatchString(a.Name) {
+		return fmt.Errorf("The 'name' field '%s' in application.yml is invalid. Only lowercase alphanumeric character(s) separated by a single hyphen are allowed. Must match regex: /^[a-z0-9]+(-[a-z0-9]+)*$/", a.Name)
+	}
+	var err error
+	serviceRoleRegex := regexp.MustCompile("^[a-zA-Z0-9]+(-[a-zA-Z0-9]+)*$")
+	a.forEachService(func(serviceType, serviceRole string, data ServiceData) {
+		if !serviceRoleRegex.MatchString(serviceRole) {
+			err = fmt.Errorf("The 'services.%s' key '%s' in application.yml is invalid. Only alphanumeric character(s) separated by a single hyphen are allowed. Must match regex: /^[a-zA-Z0-9]+(-[a-zA-Z0-9]+)*$/", serviceType, serviceRole)
+		}
+	})
+	return err
 }