+++ src/types/app_context.go
@@ -8,7 +8,6 @@ import (
 
 // AppContext represents the exosphere application the user is running
 type AppContext struct {
-	Name                     string
 	Location                 string
 	Config                   AppConfig
 	DockerComposeProjectName string
@@ -28,7 +27,6 @@ func GetAppContext(location string) (AppContext, error) {
 				return AppContext{}, err
 			}
 			return AppContext{
-				Name:     path.Base(walkDir),
 				Location: walkDir,
 				Config:   appConfig,
 			}, nil